make label for first fieldset clickable
This wasn't clickable because Typo3 inserts an empty <input> element for some reason,
thus making the <label> tag assigned to its <input> rather than the correct one.

Work around this by using the <label for=""> syntax and assign IDs to the fieldset checkboxes using their index in the array.

Change the classnames of the fieldsets to use the index as well to simplify them.