support English GOK names from 044F $a
The amount of files needing editing to achieve this seems to be deliberately high

To make this work, the ISO 639-1 language code has to be passed around to eID, which is at odds with the previous style of passing the numeric (and arbitrary) Typo3 language ID.
The current language code cannot be read from $GLOBALS['TSFE'] as it isn't initialised in eID. (Did anybody check the eID actually improves performance here? It looks like the eID get.php file loads/subclasses a lot of stuff.)
Perhaps the previous way of passing around the language code can be removed? It looks like it's not actually used.